PhuongTrinhBac2

+++ MD2/OOP/src/Rectangle/Rectangle.java
@@ -12,14 +12,16 @@ public class Rectangle {
     }
 
     public double getArea() {
+
         return this.width * this.height;
     }
 
     public double getPerimeter() {
+
         return (this.width + this.height) * 2;
     }
 
     public String display() {
-        return "Rectangle{" + "width = " + width + ",height = " + height + "}";
+        return "Rectangle{" + "width = " + (int) width + ",height = " + (int) height + "}";
     }
 }
